+++ dist-electron/preload.js
@@ -11,7 +11,16 @@ const api = {
     // Returns currently logged-in user
     GetCurrentUser: () => electron_1.ipcRenderer.invoke('get-curr-user'),
     // Returns all projects for given username
-    GetProjects: (username) => electron_1.ipcRenderer.invoke('get-user-projects', { username })
+    GetProjects: (username) => electron_1.ipcRenderer.invoke('get-user-projects', { username }),
+    // Returns all widgets for given project id
+    GetWidgets: (projectId) => electron_1.ipcRenderer.invoke('get-project-widgets', { projectId }),
+    // Creates a widget
+    CreateWidget: (projectId, name, layout) => {
+        return electron_1.ipcRenderer.invoke('create-widget', { projectId, name, layout });
+    },
+    UpdateAllWidgetLayouts: (grid) => {
+        return electron_1.ipcRenderer.invoke('update-all-widget-layouts', { grid });
+    }
 };
 electron_1.contextBridge.exposeInMainWorld('api', api);
 //# sourceMappingURL=preload.js.map