+++ electron/preload.ts
@@ -1,4 +1,13 @@
 import { contextBridge, ipcRenderer } from 'electron';
+import { UpdateAllWidgetLayouts } from './db';
+
+type widget_layout = {
+    widget_id?: number;
+    x: number;
+    y: number;
+    w: number;
+    h: number;
+}
 
 const api = {
     // Returns all users
@@ -10,6 +19,15 @@ const api = {
     // Returns currently logged-in user
     GetCurrentUser: () => ipcRenderer.invoke('get-curr-user'),
     // Returns all projects for given username
-    GetProjects: (username : string) => ipcRenderer.invoke('get-user-projects', { username })
+    GetProjects: (username : string) => ipcRenderer.invoke('get-user-projects', { username }),
+    // Returns all widgets for given project id
+    GetWidgets: (projectId : number) => ipcRenderer.invoke('get-project-widgets', { projectId }),
+    // Creates a widget
+    CreateWidget: (projectId : number, name : string, layout : widget_layout) => {
+        return ipcRenderer.invoke('create-widget', { projectId, name, layout });
+    },
+    UpdateAllWidgetLayouts: (grid : widget_layout[]) => {
+        return ipcRenderer.invoke('update-all-widget-layouts', { grid });
+    }
 }
 contextBridge.exposeInMainWorld('api', api);