+++ react/src/global.d.ts
@@ -21,6 +21,19 @@ declare global {
         username: string;
         name: string;
     };
+    type widget = {
+        id: number;
+        project_id: number;
+        name: string;
+        layout?: widget_layout;
+    }
+    type widget_layout = {
+        widget_id?: number;
+        x: number;
+        y: number;
+        w: number;
+        h: number;
+    }
     // See preload.ts for function definitions and documentation
     interface Window {
     api: {
@@ -28,6 +41,9 @@ declare global {
         LoginUser: (username : string, password : string) => Promise<response>;
         GetCurrentUser: () => Promise<response>;
         GetProjects: (username : string) => Promise<response>;
+        GetWidgets: (projectId : number) => Promise<response>;
+        CreateWidget: (projectId : number, name : string, layout : widget_layout) => Promise<response>;
+        UpdateAllWidgetLayouts: (grid : widget_layout[]) => Promise<response>;
     };
     };
 }